+++ packages/datastore-core/CHANGELOG.md
@@ -1,3 +1,33 @@
+## datastore-core [9.2.9](https://github.com/ipfs/js-stores/compare/datastore-core-9.2.8...datastore-core-9.2.9) (2024-02-28)
+
+
+### Bug Fixes
+
+* stop namespaced datastore throwing when queried ([#296](https://github.com/ipfs/js-stores/issues/296)) ([9163490](https://github.com/ipfs/js-stores/commit/916349054a3f83a7c9c4bb692edebcce409c7fee)), closes [#236](https://github.com/ipfs/js-stores/issues/236) [#236](https://github.com/ipfs/js-stores/issues/236)
+
+## datastore-core [9.2.8](https://github.com/ipfs/js-stores/compare/datastore-core-v9.2.7...datastore-core-9.2.8) (2024-02-12)
+
+
+### Bug Fixes
+
+* bump aegir to 42.2.3, update project config and fix deps ([#297](https://github.com/ipfs/js-stores/issues/297)) ([d521ef2](https://github.com/ipfs/js-stores/commit/d521ef251815527baee0a70705f775c0e47481ad))
+
+
+
+### Dependencies
+
+* **interface-datastore:** upgraded to 8.2.11
+* **interface-store:** upgraded to 5.1.8
+* **interface-datastore-tests:** upgraded to 5.1.8
+
+## [datastore-core-v9.2.7](https://github.com/ipfs/js-stores/compare/datastore-core-v9.2.6...datastore-core-v9.2.7) (2023-12-30)
+
+
+### Bug Fixes
+
+* publish with limited concurrency ([85bcc4a](https://github.com/ipfs/js-stores/commit/85bcc4acc09d76d7938c55163c81d9b948c53803))
+* readme update ([3bcfb6d](https://github.com/ipfs/js-stores/commit/3bcfb6d311d32a00f24c64cb55c3ba90ca495dba))
+
 ## [datastore-core-v9.2.6](https://github.com/ipfs/js-stores/compare/datastore-core-v9.2.5...datastore-core-v9.2.6) (2023-12-08)