+++ packages/datastore-idb/CHANGELOG.md
@@ -1,3 +1,50 @@
+## datastore-idb [2.1.9](https://github.com/ipfs/js-stores/compare/datastore-idb-2.1.8...datastore-idb-2.1.9) (2024-04-09)
+
+
+### Bug Fixes
+
+* throw read error on read error ([#304](https://github.com/ipfs/js-stores/issues/304)) ([f14c824](https://github.com/ipfs/js-stores/commit/f14c8249dfd7bbd1385bbf3513b2b3d5e0e6860c)), closes [#299](https://github.com/ipfs/js-stores/issues/299)
+
+## datastore-idb [2.1.8](https://github.com/ipfs/js-stores/compare/datastore-idb-v2.1.7...datastore-idb-2.1.8) (2024-02-12)
+
+
+### Bug Fixes
+
+* bump aegir to 42.2.3, update project config and fix deps ([#297](https://github.com/ipfs/js-stores/issues/297)) ([d521ef2](https://github.com/ipfs/js-stores/commit/d521ef251815527baee0a70705f775c0e47481ad))
+
+
+
+### Dependencies
+
+* **datastore-core:** upgraded to 9.2.8
+* **interface-datastore:** upgraded to 8.2.11
+* **interface-datastore-tests:** upgraded to 5.1.8
+
+## [datastore-idb-v2.1.7](https://github.com/ipfs/js-stores/compare/datastore-idb-v2.1.6...datastore-idb-v2.1.7) (2023-12-30)
+
+
+### Bug Fixes
+
+* publish with limited concurrency ([85bcc4a](https://github.com/ipfs/js-stores/commit/85bcc4acc09d76d7938c55163c81d9b948c53803))
+* readme update ([3bcfb6d](https://github.com/ipfs/js-stores/commit/3bcfb6d311d32a00f24c64cb55c3ba90ca495dba))
+
+
+### Trivial Changes
+
+* update sibling dependencies ([a4fb1c5](https://github.com/ipfs/js-stores/commit/a4fb1c5f97650d6ee80084e8c59c7a081f9a09e0))
+
+
+### Dependencies
+
+* bump idb from 7.1.1 to 8.0.0 ([#281](https://github.com/ipfs/js-stores/issues/281)) ([4d0bdbc](https://github.com/ipfs/js-stores/commit/4d0bdbc600b226c489259e5100af5c8c7031fb79))
+
+## [datastore-idb-v2.1.7](https://github.com/ipfs/js-stores/compare/datastore-idb-v2.1.6...datastore-idb-v2.1.7) (2023-12-28)
+
+
+### Dependencies
+
+* bump idb from 7.1.1 to 8.0.0 ([#281](https://github.com/ipfs/js-stores/issues/281)) ([4d0bdbc](https://github.com/ipfs/js-stores/commit/4d0bdbc600b226c489259e5100af5c8c7031fb79))
+
 ## [datastore-idb-v2.1.6](https://github.com/ipfs/js-stores/compare/datastore-idb-v2.1.5...datastore-idb-v2.1.6) (2023-12-01)