+++ packages/interface-blockstore-tests/CHANGELOG.md
@@ -1,3 +1,31 @@
+## interface-blockstore-tests [6.1.10](https://github.com/ipfs/js-stores/compare/interface-blockstore-tests-v6.1.9...interface-blockstore-tests-6.1.10) (2024-02-12)
+
+
+### Bug Fixes
+
+* bump aegir to 42.2.3, update project config and fix deps ([#297](https://github.com/ipfs/js-stores/issues/297)) ([d521ef2](https://github.com/ipfs/js-stores/commit/d521ef251815527baee0a70705f775c0e47481ad))
+
+
+
+### Dependencies
+
+* **interface-blockstore:** upgraded to 5.2.10
+
+## [interface-blockstore-tests-v6.1.9](https://github.com/ipfs/js-stores/compare/interface-blockstore-tests-v6.1.8...interface-blockstore-tests-v6.1.9) (2023-12-30)
+
+
+### Bug Fixes
+
+* publish with limited concurrency ([85bcc4a](https://github.com/ipfs/js-stores/commit/85bcc4acc09d76d7938c55163c81d9b948c53803))
+* readme update ([3bcfb6d](https://github.com/ipfs/js-stores/commit/3bcfb6d311d32a00f24c64cb55c3ba90ca495dba))
+
+## [interface-blockstore-tests-v6.1.8](https://github.com/ipfs/js-stores/compare/interface-blockstore-tests-v6.1.7...interface-blockstore-tests-v6.1.8) (2023-12-28)
+
+
+### Dependencies
+
+* bump multiformats from 12.1.3 to 13.0.0 ([#286](https://github.com/ipfs/js-stores/issues/286)) ([c8ccd1d](https://github.com/ipfs/js-stores/commit/c8ccd1de91883d1a1cbd394c21a51b021d52baa3))
+
 ## [interface-blockstore-tests-v6.1.7](https://github.com/ipfs/js-stores/compare/interface-blockstore-tests-v6.1.6...interface-blockstore-tests-v6.1.7) (2023-12-08)