+++ packages/interface-store/CHANGELOG.md
@@ -1,3 +1,24 @@
+## interface-store [5.1.8](https://github.com/ipfs/js-stores/compare/interface-store-v5.1.7...interface-store-5.1.8) (2024-02-12)
+
+
+### Bug Fixes
+
+* bump aegir to 42.2.3, update project config and fix deps ([#297](https://github.com/ipfs/js-stores/issues/297)) ([d521ef2](https://github.com/ipfs/js-stores/commit/d521ef251815527baee0a70705f775c0e47481ad))
+
+## [interface-store-v5.1.7](https://github.com/ipfs/js-stores/compare/interface-store-v5.1.6...interface-store-v5.1.7) (2023-12-30)
+
+
+### Bug Fixes
+
+* publish with limited concurrency ([85bcc4a](https://github.com/ipfs/js-stores/commit/85bcc4acc09d76d7938c55163c81d9b948c53803))
+
+## [interface-store-v5.1.6](https://github.com/ipfs/js-stores/compare/interface-store-v5.1.5...interface-store-v5.1.6) (2023-12-30)
+
+
+### Bug Fixes
+
+* readme update ([3bcfb6d](https://github.com/ipfs/js-stores/commit/3bcfb6d311d32a00f24c64cb55c3ba90ca495dba))
+
 ## [interface-store-v5.1.5](https://github.com/ipfs/js-stores/compare/interface-store-v5.1.4...interface-store-v5.1.5) (2023-11-27)